initial plugins work
