+++ cli/Cargo.toml
@@ -19,7 +19,7 @@
 name = "bicycle_cli"
 version = "0.1.0"
 edition = "2021"
-license-file = "../LICENSE"
+license = "AGPL-3.0"
 authors = ["sean watters sean@watters.io"]
 description = "CLI tool for the Bicycle project"
 repository = "https://github.com/ordinarylabs/bicycle"
