+++ plugin/build.rs
@@ -23,7 +23,7 @@ fn main() {
     let precompile_dir = PathBuf::from(env::var("OUT_DIR").unwrap());
 
     tonic_build::configure()
-        .file_descriptor_set_path(precompile_dir.join("bicycle_descriptor.bin"))
-        .compile(&["bicycle.proto"], &["."])
+        .file_descriptor_set_path(precompile_dir.join("plugin_descriptor.bin"))
+        .compile(&["plugin.proto"], &["."])
         .unwrap_or_else(|e| panic!("Failed to compile protos {:?}", e));
 }
